Elevator "Functioning" Subsystem
The elevator will retrieve a message from a Message object that the elevator and scheduler share. The Elevator will replace the message with an empty string (“”) and will continuously check for a valid message

“UP” message: Elevator will move up until it receives the “STOP” message
“DOWN” message: Elevator will move until it receives the “STOP” message
“STOP” message: Elevator will stop at current floor
“OPENDOORS” message: Elevator will open its doors (this takes 3 seconds)
“CLOSEDOORS” message: Elevator will open its doors (this takes 3 seconds)
“CONTINUE” message: (treat this as default)
If in IDLE state: do nothing
Else If in UP state: move up
Else If in DOWN state: move down
Every Time the elevator makes a movement between floors, the ArrivalSensor of the elevator will send a message to be retrieved by the scheduler with a single digit String which consists of the floor number e.g. “1” or “2”. With this the scheduler can send the appropriate message back for the Elevator.

No changes have been made to scheduler yet....Must be passed same messenger as elevator within main